requirements.txt
astroid==2.4.2
backcall==0.2.0
cachelib==0.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cs50==6.0.1
decorator==4.4.2
Flask==1.1.2
Flask-Session==0.3.2
gunicorn==20.0.4
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jupyter-client==6.1.7
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
Pygments==2.7.1
pylint==2.6.0
pywin32==228
pyzmq==19.0.2
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.1
tornado==6.0.4
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
